refactor: group provider modules under providers package
