+++ provider.py
@@ -1,38 +1,10 @@
-"""Provider protocol for activity sources.
+"""Compatibility shim for provider contracts.
 
-Defines the contract that all activity providers (Strava, GPX, …) must
-implement, as well as the common error type they should raise.
+Prefer importing provider contracts from ``qfit.providers.domain``.
+This module stays in place temporarily to keep older imports working while the
+feature-oriented package layout settles.
 """
 
-from typing import Dict, List, Optional, Protocol, runtime_checkable
+from .providers.domain.provider import ActivityProvider, ProviderError
 
-from .activities.domain.models import Activity
-
-
-class ProviderError(RuntimeError):
-    """Raised by an :class:`ActivityProvider` when a fetch or auth operation fails."""
-
-
-@runtime_checkable
-class ActivityProvider(Protocol):
-    """Protocol for objects that can fetch fitness activities.
-
-    Any class that exposes ``source_name``, ``last_stream_enrichment_stats``,
-    ``last_rate_limit``, and :meth:`fetch_activities` with the matching
-    signature satisfies this protocol — no explicit inheritance required.
-    """
-
-    source_name: str
-    last_stream_enrichment_stats: Dict
-    last_rate_limit: Optional[Dict]
-
-    def fetch_activities(
-        self,
-        per_page: int = 200,
-        max_pages: int = 0,
-        before: Optional[float] = None,
-        after: Optional[float] = None,
-        use_detailed_streams: bool = False,
-        max_detailed_activities: Optional[int] = None,
-    ) -> List[Activity]:
-        ...
+__all__ = ["ActivityProvider", "ProviderError"]
