+++ providers/__init__.py
@@ -1,0 +1,12 @@
+"""Provider feature package."""
+
+from .domain import ActivityProvider, ProviderError
+from .infrastructure import StravaClient, StravaClientError, StravaProvider
+
+__all__ = [
+    "ActivityProvider",
+    "ProviderError",
+    "StravaClient",
+    "StravaClientError",
+    "StravaProvider",
+]
