+++ providers/domain/__init__.py
@@ -1,0 +1,3 @@
+"""Provider domain contracts."""
+
+from .provider import ActivityProvider, ProviderError
