+++ providers/infrastructure/__init__.py
@@ -1,0 +1,4 @@
+"""Provider infrastructure adapters."""
+
+from .strava_client import StravaClient, StravaClientError
+from .strava_provider import StravaProvider
